refactor: create TranspilerService, replace DiagramService.generateDiagram() with TranspilerService.objectToMermaid()
