+++ src/app/services/transpiler/transpiler.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { TranspilerService } from './transpiler.service';
+
+describe('TranspilerService', () => {
+  let service: TranspilerService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(TranspilerService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
